+++ I2Hook/code/dcf2.h
@@ -68,6 +68,6 @@ namespace DCF2Hooks {
 
 
 	int64  HookSetProperty(int64 ptr, char* name, int64 unk);
-	void   HookReadPropertyValue(int64 ptr, int64* unk, int64* value);
+	void   HookReadPropertyValue(int64 ptr, int* unk, int* value);
 
 }
